backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dnspython==2.4.2
et-xmlfile==1.1.0
Flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
innovation-stk @ git+https://github.com/Fridaplatform/SofttekLLMSDK.git@0f11c8ed07889b9e13a360509e81b4703e03bc0f
innovation-stk==0.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.25.2
openai==0.28.0
openpyxl==3.1.2
pandas==2.1.0
pandas-stubs==2.0.3.230814
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.0.1
pinecone-client==2.2.4
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pypdf==3.16.1
PyPDF2==3.0.1
pypdfium2==4.20.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tqdm==4.66.1
types-pytz==2023.3.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.7
win32-setctime==1.1.0
yarl==1.9.2
zipp==3.17.0